add track back to index file
