+++ src/utils/auth.ts
@@ -1,5 +1,5 @@
 import OAuthClient from 'client-oauth2'
-import {track, identify} from './analytics'
+import {track, identify} from 'utils/analytics'
 import axios from 'axios'
 import get from 'lodash/get'
 import cookie from './cookies'
